fetch genres from search api call and display in search results

+++ app/src/main/java/com/lairofpixies/whatmovienext/models/network/data/TmdbMovieBasic.kt
@@ -34,5 +34,5 @@ data class TmdbMovieBasic(
     @Json(name = "poster_path")
     val posterPath: String? = null,
     @Json(name = "genre_ids")
-    val genreIds: List<Int> = emptyList(),
+    val genreIds: List<Long> = emptyList(),
 )
